removing hooks
